Refactor test_htmx_request_headers.py. Create conftest.py. Create tests for HTMX response client redirect and HTMX response stop polling. Add HTMXResponseClientRedirect and HTMXResponseStopPolling.
